Move Remove to common
