+++ samples/counter/apps/src/jvmMain/kotlin/com/slack/circuit/sample/counter/desktop/DesktopCounterCircuit.kt
@@ -38,6 +38,7 @@ import com.slack.circuit.runtime.ui.ui
 import com.slack.circuit.sample.counter.CounterPresenterFactory
 import com.slack.circuit.sample.counter.CounterScreen
 import com.slack.circuit.sample.counter.PrimeScreen
+import com.slack.circuit.sample.counter.Remove
 import kotlinx.collections.immutable.persistentListOf
 
 object DesktopCounterScreen : CounterScreen
